fixing bugs and visual bugs in purchase crypto flow, working on sell crypto flow

+++ data/src/main/java/com/multimoney/data/networking/GraphqlApi.kt
@@ -1652,7 +1652,7 @@ class GraphqlApi @Inject constructor(
         market: String,
         commissionPercentage: Double,
         taxPercentage: Double,
-        accountToken: Double,
+        accountToken: Long,
         exchangeRate: Double,
         idBrand: Int,
         user: String,
