+++ domain/src/main/java/com/multimoney/domain/interaction/crypto/SellCryptoCurrencyUseCase.kt
@@ -11,7 +11,7 @@ interface SellCryptoCurrencyUseCase {
         market: String,
         commissionPercentage: Double,
         taxPercentage: Double,
-        accountToken: Double,
+        accountToken: Long,
         exchangeRate: Double,
         idBrand: Int,
         user: String,
